Move fuzz test into a build-tag protected file
